Widen GameAccountId from int to long across plugin and server DTOs

+++ Memoria/API/Models/Requests/User/UserLoginModel.cs
@@ -8,7 +8,7 @@ namespace Memoria.API.Models.Requests.User
 {
     public class UserLoginModel
     {
-        public int GameAccountId { get; set; }
+        public long GameAccountId { get; set; }
         public string Password { get; set; } = null!;
     }
 }
